Remove warnings & add tooltips
