+++ src/main.ts
@@ -1,10 +1,11 @@
 import 'jquery';
-import 'popper.js';
-import 'bootstrap';
+
+import BootstrapVue from 'bootstrap-vue';
 import Vue from 'vue';
 import App from './App.vue';
 import store from './store';
 
+Vue.use(BootstrapVue);
 Vue.config.productionTip = false;
 
 new Vue({
