+++ src/stylesheets/frontend.scss
@@ -24,3 +24,4 @@
 // $display6-size: 1.5rem;
 
 @import "~bootstrap/scss/bootstrap.scss";
+@import "~bootstrap-vue/dist/bootstrap-vue.css";
